added background music and sound effects

+++ src/tile/TileManager.java
@@ -89,8 +89,8 @@ public class TileManager {
 
             int worldX=worldCol*gp.tileSize;
             int worldY=worldRow*gp.tileSize;
-            int screenX=(int)(worldX-gp.player.worldX+gp.player.screenX);
-            int screenY=(int)(worldY-gp.player.worldY+gp.player.screenY);
+            int screenX=worldX-gp.player.worldX+gp.player.screenX;
+            int screenY=worldY-gp.player.worldY+gp.player.screenY;
 
             g2.drawImage(tile[tileNum].image,screenX,screenY,gp.tileSize,gp.tileSize,null);
             worldCol++;
